requirements.txt
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
Django==4.0.4
djangorestframework==3.13.1
dnspython==2.2.1
mongoengine==0.24.1
numpy==1.22.3
pandas==1.4.2
pycparser==2.21
PyJWT==2.3.0
pymongo==4.1.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
